Removed special treatment for MC data in PPV vertex finder
Removed mIsMC flag in order to treat the simulated and real data in the same
manner during vertex reconstruction. The flag caused confusion rather than being
helpful in any way for the main VF task. As an additional benefit the users will
not have to worry about setting the flag when running on simulated data.